Split Status Filter tests to fix coverage OOM
Move the Status Filter describe block from TrusteesList.test.tsx
into a new TrusteesListStatusFilter.test.tsx file. Also add a
getCourts mock to the outer beforeEach in TrusteesList.test.tsx.

The large test file was causing V8 coverage instrumentation to
exhaust worker memory in GHA, resulting in 0% branch coverage for
TrusteesList.tsx and failing the 90% branch threshold. Splitting
into two files means two separate workers, each with lower peak
memory during coverage collection.

Jira ticket: CAMS-767
